backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
supabase==2.7.4
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
openai==1.51.0
anthropic==0.34.2
deepgram-sdk==3.7.7
pdfplumber==0.11.4
mammoth==1.8.0
mercadopago==2.3.0
